refactor: 인증 위치 변경
