+++ server/src/main/java/com/oddok/server/domain/user/mapper/AuthMapper.java
@@ -1,6 +1,7 @@
 package com.oddok.server.domain.user.mapper;
 
 import com.oddok.server.domain.user.api.response.AuthResponse;
+import com.oddok.server.domain.user.api.response.UpdateTokenResponse;
 import com.oddok.server.domain.user.dto.TokenDto;
 import com.oddok.server.domain.user.dto.TokensDto;
 import org.mapstruct.Mapper;
@@ -9,4 +10,5 @@ import org.mapstruct.Mapping;
 @Mapper
 public interface AuthMapper {
     AuthResponse toAuthResponse(TokensDto dto);
+    UpdateTokenResponse toTokenResponse(TokenDto dto);
 }
