+++ server/src/main/java/com/oddok/server/domain/user/mapper/UserDtoMapper.java
@@ -11,6 +11,4 @@ public interface UserDtoMapper {
 
     ChangeNicknameResponse toChangeNicknameResponse(UserDto dto);
 
-    UpdateTokenResponse toTokenResponse(TokenDto dto);
-
 }
